refactor: UserController 코드 정리
ListUserController의 user 목록 정보를 가져오는 메서드를 UserController로 이동하였습니다.
