--- slipp/src/main/java/slipp/controller/ListUserController.java
@@ -1,25 +1,0 @@
-package slipp.controller;
-
-import nextstep.mvc.tobe.ModelAndView;
-import nextstep.mvc.tobe.view.JspView;
-import nextstep.mvc.tobe.view.RedirectView;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.annotation.RequestMethod;
-import slipp.support.db.DataBase;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@Controller
-public class ListUserController {
-    @RequestMapping(value = "/users/list", method = RequestMethod.GET)
-    public ModelAndView userList(HttpServletRequest req, HttpServletResponse resp) {
-        if (!UserSessionUtils.isLogined(req.getSession())) {
-            return new ModelAndView(new RedirectView("/users/loginForm"));
-        }
-
-        req.setAttribute("users", DataBase.findAll());
-        return new ModelAndView(new JspView("/user/list.jsp"));
-    }
-}
