+++ slipp/src/main/java/slipp/controller/UserController.java
@@ -6,8 +6,7 @@ import nextstep.mvc.tobe.view.RedirectView;
 import nextstep.web.annotation.Controller;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.annotation.RequestMethod;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import slipp.controller.exception.UserNotFoundException;
 import slipp.domain.User;
 import slipp.support.db.DataBase;
 
@@ -16,44 +15,62 @@ import javax.servlet.http.HttpServletResponse;
 
 @Controller
 public class UserController {
-    private static final Logger log = LoggerFactory.getLogger(UserController.class);
-
     @RequestMapping(value = "/users/create", method = RequestMethod.POST)
     public ModelAndView create(HttpServletRequest req, HttpServletResponse resp) {
-        User user = new User(
-                req.getParameter("userId"), req.getParameter("password"),
-                req.getParameter("name"), req.getParameter("email")
-        );
-        log.debug("User : {}", user);
+        User user = createUserByRequestBody(req);
 
         DataBase.addUser(user);
         return new ModelAndView(new RedirectView("/"));
+    }
+
+    @RequestMapping(value = "/users/list", method = RequestMethod.GET)
+    public ModelAndView userList(HttpServletRequest req, HttpServletResponse resp) {
+        if (!UserSessionUtils.isLogined(req.getSession())) {
+            return new ModelAndView(new RedirectView("/users/loginForm"));
+        }
+
+        req.setAttribute("users", DataBase.findAll());
+        return new ModelAndView(new JspView("/user/list.jsp"));
     }
 
     @RequestMapping(value = "/users/profile", method = RequestMethod.GET)
     public ModelAndView find(HttpServletRequest req, HttpServletResponse resp) {
         String userId = req.getParameter("userId");
         User user = DataBase.findUserById(userId);
-        if (user == null) {
-            throw new NullPointerException("사용자를 찾을 수 없습니다.");
-        }
+
+        checkExisting(user);
         req.setAttribute("user", user);
         return new ModelAndView(new JspView("/user/profile.jsp"));
     }
 
     @RequestMapping(value = "/users/update", method = RequestMethod.POST)
     public ModelAndView update(HttpServletRequest req, HttpServletResponse resp) {
-        User user = DataBase.findUserById(req.getParameter("userId"));
-        if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
-            throw new IllegalStateException("다른 사용자의 정보를 수정할 수 없습니다.");
-        }
+        User user = getAuthenticatedUser(req);
+        User updatedUser = createUserByRequestBody(req);
 
-        User updateUser = new User(
+        user.update(updatedUser);
+        return new ModelAndView(new RedirectView("redirect:/"));
+    }
+
+    private User createUserByRequestBody(HttpServletRequest req) {
+        return new User(
                 req.getParameter("userId"), req.getParameter("password"),
                 req.getParameter("name"), req.getParameter("email")
         );
-        log.debug("Update User : {}", updateUser);
-        user.update(updateUser);
-        return new ModelAndView(new RedirectView("redirect:/"));
+    }
+
+    private User getAuthenticatedUser(HttpServletRequest req) {
+        User user = DataBase.findUserById(req.getParameter("userId"));
+
+        if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
+            throw new IllegalStateException("다른 사용자의 정보를 수정할 수 없습니다.");
+        }
+        return user;
+    }
+
+    private void checkExisting(User user) {
+        if (user == null) {
+            throw new UserNotFoundException("사용자를 찾을 수 없습니다.");
+        }
     }
 }
